+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 # [8.10.0] - 2024-08-??
 - Added `targetApiKey` for buy & cancel number endpoints
-- 
+- Added Builder for `UpdateNumberRequest` and deprecated setters
+- Added `app_id` property to `UpdateNumberRequest` (replaces `messagesCallbackValue`)
 
 # [8.9.4] - 2024-07-25
 - Fixed UUID validation in `ConversationAction.Builder#canHear` and `canSpeak`